third_party/blink/web_tests/flag-specific/enable-blink-features=ForceSynchronousHTMLParsing/external/wpt/custom-elements/perform-microtask-checkpoint-before-construction-expected.txt

This is a testharness.js-based test.
PASS HTML parser must perform a microtask checkpoint before constructing a custom element
FAIL HTML parser must perform a microtask checkpoint before constructing a custom element during the adoption agency algorithm assert_equals: expected 1 but got 2
Harness: the test ran to completion.
